Extract StreamDiscoveryContext to executor/stream_discovery.rs module
Third small step in executor.rs refactoring. This change moves the
StreamDiscoveryContext type and its implementation to a dedicated
module while maintaining all existing APIs and functionality.

Also updates all field accesses to use proper getter methods for
better encapsulation. All tests pass with no behavioral changes.
